Renamed to remove impl keyword
