+++ conf/script/build_system/compiler/host/get_info/version/msvc/cli.py
@@ -4,7 +4,7 @@ from typing import NoReturn, Optional, Union
 from cli_fetch_compiler_info import cli_fetch_compiler_info
 from version import CompilerVersion
 from data_model import Compiler
-from fetch_msvc_version_impl import fetch_msvc_version
+from fetch_msvc_version import fetch_msvc_version
 from cli_find_msvc_location import error_compiler_not_found
